+++ tests/testthat/test-req-body.R
@@ -41,6 +41,20 @@ test_that("can send any type of object as json", {
   expect_equal(json$json, as.list(letters))
 })
 
+test_that("can use custom json type", {
+  req <- request_test("/post") %>% req_body_json(mtcars)
+  expect_equal(req_body_apply(req)$headers$`Content-Type`, "application/json")
+
+  content_type <- "application/ld+json"
+  req <- req |> req_headers(`Content-Type` = "application/ld+json")
+  expect_equal(req_body_apply(req)$headers$`Content-Type`, content_type)
+
+  expect_snapshot({
+    (expect_error(req |> req_headers(`Content-Type` = "application/ld+json2") |> req_body_apply()))
+    (expect_error(req |> req_headers(`Content-Type` = "app/ld+json") |> req_body_apply()))
+  })
+})
+
 test_that("can send named elements as form/multipart", {
   data <- list(a = "1", b = "2")